feat(iota-indexer-streaming): isolate stream module form `iota-indexer` into separate crate (#8949)
# Description of change

This PR isolates the `stream` module in `iota-indexer` into a separate
crate `iota-indexer-streaming`. This will allow us to evolve and
maintain it independently.

## Links to any relevant issues

fixes #8839

## How the change has been tested

- [x] Basic tests (linting, compilation, formatting, unit/integration
tests)
- [ ] Patch-specific tests (correctness, functionality coverage)

- started a local network with graphql server using the `iota` cli tool
to assert that the streaming was working as expected.

### Infrastructure QA (only required for crates that are maintained by
@iotaledger/infrastructure)

- [ ] Synchronization of the indexer from genesis for a network
including migration objects.
- [ ] Restart of indexer synchronization locally without resetting the
database.
- [ ] Restart of indexer synchronization on a production-like database.
- [ ] Deployment of services using Docker.
- [ ] Verification of API backward compatibility.

> [!NOTE]
> The patch does not affect the indexer normal operations, thus no tests
were conducted.

### Release Notes

- [x] Indexer: introduce to monorepo the `iota-indexer-streaming` crate
which holds logic related to `iota-indexer` streaming capabilities.
