+++ crates/iota-graphql-rpc/Cargo.toml
@@ -55,6 +55,7 @@ iota-graphql-config.workspace = true
 iota-graphql-rpc-client.workspace = true
 iota-graphql-rpc-headers.workspace = true
 iota-indexer.workspace = true
+iota-indexer-streaming.workspace = true
 iota-json.workspace = true
 iota-json-rpc.workspace = true
 iota-json-rpc-api.workspace = true
