+++ crates/iota-graphql-rpc/src/types/subscription/filter.rs
@@ -2,9 +2,9 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use async_graphql::OneofObject;
-use iota_indexer::stream::{ModuleFunction, StreamEventFilter, StreamTransactionFilter};
-use iota_json_rpc_types::IotaTransactionKind;
-use iota_types::base_types::ObjectID;
+use iota_indexer::models::{events::StoredEvent, transactions::StoredTransaction};
+use iota_json_rpc_types::{Filter, IotaTransactionKind};
+use iota_types::{base_types::ObjectID, transaction::TransactionDataAPI};
 
 use crate::types::{
     iota_address::IotaAddress,
@@ -22,18 +22,15 @@ pub(crate) enum SubscriptionEventFilter {
     EmittingModule(ModuleFilter),
 }
 
-impl From<SubscriptionEventFilter> for StreamEventFilter {
-    fn from(value: SubscriptionEventFilter) -> Self {
+impl Filter<StoredEvent> for SubscriptionEventFilter {
+    fn matches(&self, event: &StoredEvent) -> bool {
         use SubscriptionEventFilter::*;
-        match value {
+        match self {
             EmittingModule(ModuleFilter::ByPackage(package)) => {
-                StreamEventFilter::EmittingPackage(package.into())
+                event.package.as_slice() == package.as_slice()
             }
             EmittingModule(ModuleFilter::ByModule(package, module)) => {
-                StreamEventFilter::EmittingModule {
-                    package: package.into(),
-                    module,
-                }
+                event.package.as_slice() == package.as_slice() && event.module == *module
             }
         }
     }
@@ -54,22 +51,34 @@ pub(crate) enum SubscriptionTransactionFilter {
     Function(FqNameFilter),
 }
 
-impl From<SubscriptionTransactionFilter> for StreamTransactionFilter {
-    fn from(value: SubscriptionTransactionFilter) -> Self {
+impl Filter<StoredTransaction> for SubscriptionTransactionFilter {
+    fn matches(&self, transaction: &StoredTransaction) -> bool {
         use SubscriptionTransactionFilter::*;
-        match value {
-            Kind(kind) => StreamTransactionFilter::Kind(kind.into()),
-            SigningAddress(address) => StreamTransactionFilter::SigningAddress(address.into()),
+        match self {
+            Kind(kind) => transaction.transaction_kind == IotaTransactionKind::from(kind) as i16,
+            SigningAddress(address) => transaction
+                .try_into_sender_signed_data()
+                .map(|data| data.transaction_data().sender() == (*address).into())
+                .unwrap_or_default(),
             Function(name) => {
-                let (package, module) = name.into();
-                StreamTransactionFilter::Function { package, module }
+                let move_call = MoveCall::from(name);
+
+                transaction
+                    .try_into_sender_signed_data()
+                    .map(|data| {
+                        data.transaction_data()
+                            .move_calls()
+                            .iter()
+                            .any(|(p, m, f)| move_call.matches_transaction_move_call(p, m, f))
+                    })
+                    .unwrap_or_default()
             }
         }
     }
 }
 
-impl From<TransactionBlockKindInput> for IotaTransactionKind {
-    fn from(value: TransactionBlockKindInput) -> Self {
+impl From<&TransactionBlockKindInput> for IotaTransactionKind {
+    fn from(value: &TransactionBlockKindInput) -> Self {
         match value {
             TransactionBlockKindInput::SystemTx => IotaTransactionKind::SystemTransaction,
             TransactionBlockKindInput::ProgrammableTx => {
@@ -90,24 +99,70 @@ impl From<TransactionBlockKindInput> for IotaTransactionKind {
     }
 }
 
-impl From<FqNameFilter> for (ObjectID, Option<ModuleFunction>) {
-    fn from(value: FqNameFilter) -> Self {
+/// Represents a module information of the move call.
+struct ModuleFunction<'a> {
+    /// Name of the module.
+    module_name: &'a str,
+    /// Name of the function within the module.
+    function_name: Option<&'a str>,
+}
+
+/// A data type that converts [`FqNameFilter`] into a representation of move
+/// calls in transactions, enabling easy filtering by fully qualified names as
+/// returned by the [`move_calls`](TransactionDataAPI::move_calls) method on
+/// types implementing [`TransactionDataAPI`].
+struct MoveCall<'a> {
+    /// Package ID of the move call.
+    package: ObjectID,
+    /// Module information of the move call.
+    module_function: Option<ModuleFunction<'a>>,
+}
+
+impl<'a> MoveCall<'a> {
+    fn new(package: ObjectID, module_function: impl Into<Option<ModuleFunction<'a>>>) -> Self {
+        MoveCall {
+            package,
+            module_function: module_function.into(),
+        }
+    }
+
+    /// Matches a transaction move call against the fully qualified name filter.
+    ///
+    /// The filter is applied in the following order:
+    /// 1. package ID must match
+    /// 2. if a module name is specified, it must match
+    /// 3. if a function name is specified, it must match
+    fn matches_transaction_move_call(
+        &self,
+        package: &ObjectID,
+        module: &str,
+        function: &str,
+    ) -> bool {
+        self.package == *package
+            && self.module_function.as_ref().is_none_or(|mf| {
+                mf.module_name == module && mf.function_name.is_none_or(|f| f == function)
+            })
+    }
+}
+
+impl<'a> From<&'a FqNameFilter> for MoveCall<'a> {
+    fn from(value: &'a FqNameFilter) -> Self {
         use FqNameFilter::*;
         match value {
-            ByModule(ModuleFilter::ByPackage(package)) => (package.into(), None),
-            ByModule(ModuleFilter::ByModule(package, module_name)) => (
-                package.into(),
-                Some(ModuleFunction {
+            ByModule(ModuleFilter::ByPackage(package)) => Self::new((*package).into(), None),
+            ByModule(ModuleFilter::ByModule(package, module_name)) => Self::new(
+                (*package).into(),
+                ModuleFunction {
                     module_name,
                     function_name: None,
-                }),
+                },
             ),
-            ByFqName(package, module_name, function_name) => (
-                package.into(),
-                Some(ModuleFunction {
+            ByFqName(package, module_name, function_name) => Self::new(
+                (*package).into(),
+                ModuleFunction {
                     module_name,
                     function_name: Some(function_name),
-                }),
+                },
             ),
         }
     }
