+++ crates/iota-indexer-streaming/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "iota-indexer-streaming"
+version.workspace = true
+authors = ["IOTA Foundation <info@iota.org>"]
+edition = "2021"
+license = "Apache-2.0"
+publish = false
+
+[dependencies]
+# external dependencies
+async-trait.workspace = true
+futures.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+thiserror.workspace = true
+tokio = { workspace = true, features = ["rt"] }
+tokio-postgres = "0.7.15"
+tokio-stream.workspace = true
+tracing.workspace = true
+
+# internal dependencies
+iota-indexer.workspace = true
+iota-types.workspace = true
