+++ crates/iota-indexer-streaming/src/lib.rs
@@ -1,0 +1,8 @@
+// Copyright (c) 2025 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+//! Provides real-time streaming of transactions and events from the IOTA
+//! Indexer.
+
+pub mod error;
+pub mod memory;
