+++ crates/iota-indexer/Cargo.toml
@@ -28,7 +28,6 @@ prometheus.workspace = true
 rand = { workspace = true, optional = true }
 secrecy = "0.8.0"
 serde.workspace = true
-serde_json.workspace = true
 serde_with.workspace = true
 strum.workspace = true
 strum_macros.workspace = true
@@ -36,7 +35,6 @@ tap.workspace = true
 tempfile.workspace = true
 thiserror.workspace = true
 tokio = { workspace = true, features = ["full"] }
-tokio-postgres = "0.7.13"
 tokio-stream.workspace = true
 tokio-util = { workspace = true, features = ["rt"] }
 toml.workspace = true
