+++ crates/iota-indexer/src/lib.rs
@@ -44,7 +44,6 @@ pub mod pruning;
 pub mod read;
 pub mod schema;
 pub mod store;
-pub mod stream;
 pub mod system_package_task;
 pub mod test_utils;
 pub mod types;
